drm/radeon: stop poisoning the GART TLB
commit 0986c1a55ca64b44ee126a2f719a6e9f28cbe0ed upstream.

When we set the valid bit on invalid GART entries they are
loaded into the TLB when an adjacent entry is loaded. This
poisons the TLB with invalid entries which are sometimes
not correctly removed on TLB flush.

For stable inclusion the patch probably needs to be modified a bit.

Change-Id: Ic8ea0600665a830efab855fcf3490d1dd4dea6fb
Signed-off-by: Christian König <christian.koenig@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>

	modified:   drivers/gpu/drm/radeon/rs600.c
Change-Id: I703d2b14969d3bde4f0ab006b8465a76edac5270